Export IgnoreMinimumTapAreaSize to library users
