+++ lib/accessibility_tools.dart
@@ -1,6 +1,7 @@
 library accessibility_tools;
 
 export 'src/accessibility_tools.dart';
+export 'src/checkers/ignore_minimum_tap_area_size.dart';
 export 'src/checkers/minimum_tap_area_checker.dart' show MinimumTapAreas;
 export 'src/enums/buttons_alignment.dart';
 export 'src/testing_tools/testing_tools_configuration.dart';
