+++ lib/src/checkers/minimum_tap_area_checker.dart
@@ -7,6 +7,7 @@ import 'package:flutter/rendering.dart';
 
 import '../accessibility_issue.dart';
 import 'checker_base.dart';
+import 'ignore_minimum_tap_area_size.dart';
 
 /// Defines the minimum tap size per device type.
 class MinimumTapAreas {
@@ -165,28 +166,5 @@ Icon(
         nodePaintBounds.left < -delta ||
         nodePaintBounds.bottom > windowPhysicalSize.height + delta ||
         nodePaintBounds.right > windowPhysicalSize.width + delta;
-  }
-}
-
-class IgnoreMinimumTapAreaSize extends InheritedWidget {
-  const IgnoreMinimumTapAreaSize({
-    super.key,
-    required super.child,
-  });
-
-  static IgnoreMinimumTapAreaSize? maybeOf(BuildContext context) {
-    return LookupBoundary.dependOnInheritedWidgetOfExactType<IgnoreMinimumTapAreaSize>(context);
-  }
-
-  static IgnoreMinimumTapAreaSize of(BuildContext context) {
-    final IgnoreMinimumTapAreaSize? result =
-        context.dependOnInheritedWidgetOfExactType<IgnoreMinimumTapAreaSize>();
-    assert(result != null, 'No IgnoreMinimumTapAreaSize found in context');
-    return result!;
-  }
-
-  @override
-  bool updateShouldNotify(IgnoreMinimumTapAreaSize oldWidget) {
-    return false;
   }
 }
